+++ lib/CGF/include/Sprite.h
@@ -61,10 +61,10 @@ public:
     bool getMirror() { return mirror; }
 
     // Sprite speed
-    void setXspeed(double xspeed);
-    void setYspeed(double yspeed);
-    double getXspeed() { return xspeed; }
-    double getYspeed() { return yspeed; }
+    void setXspeed(float xspeed);
+    void setYspeed(float yspeed);
+    float getXspeed() { return xspeed; }
+    float getYspeed() { return yspeed; }
 
     // Animation control
     void setAnimation(std::string name);
@@ -90,7 +90,7 @@ public:
 
     sf::Vector2u getSize() { return sf::Vector2u(spriteW, spriteH); }
 
-    void update(double deltaTime, bool updatePos=true);
+    void update(float deltaTime, bool updatePos = true);
 
     // Basic collision checking
     bool bboxCollision(Sprite& other);
@@ -123,7 +123,7 @@ private:
     bool mirror;
 
     // Motion
-    double xspeed,yspeed;       // speed in pixels/s
+    float xspeed,yspeed;       // speed in pixels/s
     int updateCount;            // current count of updates
 
     // Animation
